requirements.txt
python-dotenv>=1.0.0
requests>=2.28.2
numpy>=1.24.3
pandas>=2.1.0
Flask>=2.3.2
Jinja2>=3.1.2
opencv-python-headless>=4.8.0
Pillow>=10.0.0
streamlit>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.10.1
matplotlib>=3.7.1
roboflow>=1.1.0
psycopg2-binary>=2.9.6
SQLAlchemy>=2.0.19
alembic>=1.10.2
gunicorn>=20.1.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0
tqdm>=4.65.0